Added the model and pages
