+++ Views/Inference/PredictedScore.cshtml
@@ -1,0 +1,11 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@model Intex2A.Models.PredictionModels.Prediction
+
+<div>
+    <h1>Prediction Result:</h1>
+    <p>@Model.PredictedValue</p>
+</div>
+
+
